pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "The tool to generate Mastodon OpenAPI spec"
authors = ["cmj <cmj@cmj.tw>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.13"
pydantic = "^2.10.4"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
